Run functional tests against Java implementation
Converts the functional tests to use JPype to run the actual Java
implementation. This will prevent accidentally adding incorrect test
expectations based on a misreading of the spec.

The tests are also run against wrappers to pyjavaproperties and
python-jproperties to compare their implementations. These tests are marked as
xfail so that failures will just be reported, but not considered a failure for
the test run.