Support text-mode files for other encodings
Converts most of the inner workings to use text strings instead of byte strings.
Moves encoding to the outer layer so that if the file appears to have been
opened with an encoding (e.g. io.TextIOBase or similar) it will rely on the
existing encoding support. Files in bytes mode will continue to use the
"latin-1" encoding as specified in the Java documentation.